Pass dirname(filePath) to require.resolve
This will make nodejs search recursively from the location of the file
being formatted. The idea is that node_modules may not be located at the
current working dir, so we leverage node's ability to require modules
and just pass the dirname of the provided filepath

Also bump the cache limits because we'll be splitting the cache a bit
more (maybe I'm bumping it too far, idk).

Closes #245.