feat: support ignoring files with .prettierignore (#275)
Add support for ignoring files specified in `.prettierignore`.

The path to the ignore file can be customized by passing in the
`--ignore-path` option, similarly to prettier CLI.
See https://prettier.io/docs/en/cli.html#--ignore-path for more
information.

The `--ignore-path` CLI flag can appear before of after the file name.
The order of CLI arguments does not matter.

This commit adds a bit more robust CLI arguments parsing. There will be
more explicit errors when the path to the file to format is not provided
(prettier expects it anyway and would throw otherwise) and when the
`--ignore-path` flag is passed, but its argument is not.

Prettier's public `getFileInfo` API is used to reuse the same logic that
Prettier uses for determining whether the file should be ignored or not.
See https://prettier.io/docs/en/api.html#prettiergetfileinfofilepath--options

The performance does not seem to have degraded. I did not notice
slowdowns with the extra call to `getFileInfo`.