Use get-stdin to read the stdin
It's not _really_ a new dependency as prettier already depends on it,
but let's bring it in.

I suspect that users are running into weird encoding issues in #694 (and
maybe #698).

Rather than spending too much time on this, I'm just taking a shortcut:
users confirm that they cannot reproduce the issue in `prettier`, so
let's read stdin the same `prettier` does :)