Simply vendor `get-stdin`
The current state of affairs in prettierd are a bit too tricky when it
comes to modules: it deeply depends on commonjs to dynamically resolve
which prettier to import, and that isn't supported out of the box with
esmodules. It turns out that get-stdin doesn't support cjs, and
typescript is being annoying here, compiling `import` into `require`.

To be honest, migrating to esmodules wouldn't be super tricky (famous
last words): we'd need to implement some of the filesystem navigation to
find the files and then import those with `import()`. The thing is, I'm
not 100% sure that that's trivial, and I allotted some time to work on
prettierd today and that time is running out, and I want to make sure I
fix at least one bug without getting nerdsnipped.

So, rather than trying to migrate everything to esmodules, let's keep it
simple and just vendor `get-stdin`.

Another option I considered was bypassing TS to still use `import()`
with commonjs (some options here:
https://github.com/TypeStrong/ts-node/discussions/1290), but that sounds
too painful.